Perform Apricot Phase 2 Validation on EVM ImportTxs (#187)
ImportTxs for the EVM now have to obey two new rules:

* The array of EVMOutputs must not have any duplicate (assetID, address) pairs. That is, there cannot be multiple outputs sending the same asset to the same address. A given address may receive multiple assets, and a given asset may be sent to multiple addresses, but you can't (for example) send my address some AVAX in two separate Outputs.
* The importing transaction must burn enough AVAX to cover a transaction fee. That is, the difference between AVAX-related Input amounts and AVAX-related EVMOutputs must be equal to or greater than the transaction fee.

At the end of ImportTx construction, validateOuts is called to enforce these rules, throwing errors if they are broken. Because this function needs access to ins and outs, but it is legal to call the constructor without ins and outs, validateOuts is only called if both ins and outs is provided to the constructor.